Fix 'embedded' Nginx instance to work on Arch Linux.
On Arch, the nginx package is configured differently than on Mac OS X,
which is where I was testing.
A lot of hardcoded paths that Nginx tries to write files to are
specified as paths owned by root.
To avoid this I added a lot more directives to the generated nginx.conf
that cause Nginx to instead use a temporary directory.